Add the relevant files from `high-assurance-legacy`
We do not add entries to `ROOT` for the various theories, since they do
not work with the current libraries, like the Þ-calculus and the
next-generation equivalence reasoner. Making them compatible requires
also changes to said libraries, at least to the Þ-calculus, which in
particular doesn’t yet have a typed layer. We shall adapt the files
added herewith later.