Widen the use of extra split rules
This change does justice to the fact that also `let` expressions with
pattern matching need the corresponding `case` split rules passed. Also
it fixes a bug in the previous documentation, which asked to add only
rules corresponding to the types of scrutinees but not to the types of
subpatterns that perform additional case distinction.