Change `;` to `.` in `Await` notation
We want to use `.` in the mini-protocol program notation for receiving
from synchronous channels, to make it consistent with the corresponding
notation on the Þ-calculus level. The notation for `Await` should use
`.` to be consistent with the mini-protocol program notation for
receiving from synchronous channels but also to be consistent with the
notation of the receive constructs of the Þ-calculus.