Improve identifier `state_in_server_phase`
For the `Idle` phase, there is not just one state that the server can be
in; so the above name is misleading. This function is about the state
that the server is in at the beginning of a cycle. We call this now the
base state and have chosen the new identifier accordingly.